zerocopy views for table scans

+++ crates/db-catalog/src/table.rs
@@ -106,7 +106,7 @@ impl DBTable {
                 .expect("table paritition exists");
 
             let row_count_chunk = (row_count_total - row_count_done)
-                .min(table_paritition.rows_available() as usize);
+                .min(table_paritition.rows_available());
             let byte_start = row_count_done * row_byte_width;
             let byte_end = byte_start + row_count_chunk * row_byte_width;
             table_paritition.insert_rows(schema, &bytes[byte_start..byte_end]);
